Coin (#23)
* Skybox Changed & SkiyLine Added

* NFt Market_Pumink added

* Minigame_mole Display & NPC SpeechBubble Added

* speech bubble edited

* Finished UI interaction on Game01

* bassic Dialogue System

* auto conversation

* final ik, avatar setting

* final ik non leg

* height solved  -> ovrplayercontroller <use profile data >false

* foot ik finished

* npc basic conversation

* Npc Textnumber saveed

* Player SpawnPoints

* main Spawnpoint added

* House relocated

* Light delete and NPC Conversation Line added

* temp

* merge test

* Change Bullings

* tutorial to main scene 이동 구현

* Light Option Edited

* remove some files

* 인형공방 디자인

* FxEffectAssets

* Lightg Option edited

* Light options Edited

* Skybox Changed & SkiyLine Added

* NFt Market_Pumink added

* Minigame_mole Display & NPC SpeechBubble Added

* working Game02 WhaAShark

* Play Game02 working

* Shark game working

* SharkUiWorking

(cherry picked from commit 6c165bb9b629d197145aa006284ca9e571966c86)

* Shark added

* Sign Added

* Sign Animations added

* Npc Conversation Edited

* birdGameWorking

* BirdBombWorkingBeforeLunch

* birdGameMiddleCheck

* birdGameWorking

* BirdGameWork

* 위치 조정

* working

* Patrol Npc adeed

* Pumkin Area NPC added

* _Assets deleted

* Paragliding game upload

* added assets & Doll WorkShop Interior

* paraglidingWorking02

* LightMap Bake 01

* Paragliding&FireworkFestivalWorking

* Map03 LightTesting

* Final BirdBombBowling Game, Paragliding Game

* correct error

* remove house, tree etc..

* 메인 씬 디자인

* 모든 공방에서 nft 오브젝트 잡아서 볼 수 있는 기능 추가 & 가격 / 이름 표시

* nft 제조 카운터생성

* 인형공방 & 음악공방 굿즈들에 rigidbody 추가
TODO : 안쓰러지게 해야함 (콜라이더 조정 필요)

* collider 조정 완료

* 맵 포톤 동기화 테스트 맵01 복사함

* 보이스챗 테스트

* 포톤챗 완

* change gitattributes

* nft object 계산대에 올려놓기 구현

* 기린 인형 제거

* photon test

* sound 거리 조정

* 도어 애니메이터

* 튜토 스킵

* 튜토 스킵

* 거울,풋애니메이터,텔레포트크기,인게임레이

* 거울,풋애니메이터,텔레포트크기,인게임레이2

* 레이포톤수정

* 레이포톤수정

* photonserialize

* music work shop design added

* 인형 공방, 음악 공방도 카운터 구현

* 라인 렌더 불

* merge

* ingame레이만

* 머지메인

* 맵 수정

* add box collider

* add reflection probe

* 음악공방 및 npc 이동오류 해결

* 아바타카메라컬링, 렌더나가는거 해결,그랩오류아직있음

* 돌 애니메이션 수정

* Finishing Bird,Paragliding Game

* scene correct

* game working paragliding

* correct

* Lobby Loading Scene & Map01 Light

* 로비 fps

* mergetest

* modify gitattributes

* apply gitattributes

* Game working

* nft object 구매 완료

* light edite

* Game in Map01

* FinalGameWorking02

* FinalGameWorkingCorrect

* bug fix

* Map_01 씬으로 안넘어가는 문제 해결중

* fps 캔버스 완

* bug fix

* Light Test

* photonray 합완

* minigame adding

* 도어 포톤뷰지움 포톤매니저 인스턴시에이터 처음으로

* Minigame added & LOD deleted

* SharkGameDone

* 캔버스 unassign null 해결 포톤 연동해서 확인해봐야함 포톤챗 고쳐야함

* 포톤챗 마이크 디바이스 변경해봄

* SharkGameWorking02

* BirdGameWorking02

* bird Game Working03

* not pvismine카메라 컬링레이어 변경

* add map_01_plus

* merge develop

* remove prefab file, json file

* assign

* temp

* 두더지 포톤연동해봄

* 두더지 포톤연동해봄 photnnetwork loadlevel

* 미니게임 인아웃 연동

* 미니게임 인아웃 연동

* coin api 연동

* 보이스매니저 포톤 -> 유니티

* coin 지갑에 연동

* coin 연동

* 티켓 위치 이상한 현상 수정

* 공방 오브젝트 코인이름, 가격 변경

* 미니게임 포톤 연동

+++ Assets/02.Scripts/NFT/CoinCache.cs
@@ -1,4 +1,4 @@
 public class CoinCache
 {
-    public static int coin;
+    public static double coin;
 }
